Implement data storage in heap-allocated arrays
Prior to this data was stored in a set of stack-allocated arrays, the size of
which was decided at compile-time.
The problems with this are that:
  - The number of samples could not be decided at run-time, which is something
    we'll want to implement later
  - If you wanted to take more than 300k samples the program would crash since
    it couldn't allocate that much space on the stack
This change makes the intermediary data storage arrays heap-allocated,
considerably expanding the storage capacity and making dynamic assignment
possible in the future.

Performance impact TBA.