Refactor into separate files
This change moves all of the mundane stuff (setting up the PRU, saving data etc)
to another set of files so the main is a bit cleaner.

The main data capture routine was not refactored out because I don't want to
deal with the overhead of a functionality, and I'm too lazy to check the output
of the compiler to see if it would be automatically inlined.