Make the timestamp array 1D instead of 2D
This was a huge blunder on my part.
The timestamp array ended up being the same size as the data array, but full of
64-bit integers... i.e. it was a huge waste of memory.

With this change, the timestamp array is simply a 1D array that's used in
parallel with the 2D data array.
This makes it possible to record huge quantities of data, I've tested it up to
10 million measurements over 6 channels without crashes.
At this point it seems that the biggest bottleneck is copying the data from
memory to the file, not reading it!